% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{AmigaBasic-files}
\alias{AmigaBasic-files}
\alias{demo.bas}
\alias{r_logo.shp}
\alias{ball.shp}
\title{'demo.bas', 'r_logo.shp' and 'ball.shp' as example files for AmigaBasic and AmigaBasicShape objects}
\format{
See \code{\link{AmigaBasic}} and \code{\link{AmigaBasicShape}} for more information
about the format.
}
\description{
TODO
}
\details{
TODO
}
\examples{
\dontrun{
TODO
}
}
