% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{AmigaBasic.names}
\alias{AmigaBasic.names}
\alias{AmigaBasic.names<-}
\title{Extract or replace variable and label names from Amiga Basic scripts}
\usage{
AmigaBasic.names(x)

AmigaBasic.names(x) <- value
}
\arguments{
\item{x}{XXX}

\item{value}{XXX}
}
\value{
XXX
}
\description{
In the binary Amiga Basic files, names for labels and variabels
in the code are stored at the end of the file. In the encoded
there is only a reference to the index of the name in that list. Use
this function to list, select or replace names included in the code
}
\details{
XXX
}
\examples{
## Let's create some Basic code with labels and variables:
bas <- as.AmigaBasic(c(
  "REM - This will loop forever...",
  "my.label:",
  "  my.variable\% = 0",
  "  WHILE my.variable\% < 10",
  "    my.variable\% = my.variable\% + 1",
  "  WEND",
  "  GOTO my.label"
))

## list the names in the script above:
AmigaBasic.names(bas)

## change the first name:
AmigaBasic.names(bas)[1] <- "better.label"
}
\seealso{
Other AmigaBasic.operations: \code{\link{AmigaBasic.reserved}},
  \code{\link{AmigaBasic}}, \code{\link{[.AmigaBasic}},
  \code{\link{as.AmigaBasic}}, \code{\link{as.character}},
  \code{\link{rawToAmigaBasic}},
  \code{\link{read.AmigaBasic}},
  \code{\link{write.AmigaBasic}}
}
\author{
Pepijn de Vries
}
