% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\docType{class}
\name{AmigaBasic}
\alias{AmigaBasic}
\title{The S3 AmigaBasic class}
\description{
XXX
}
\details{
XXX add to package documentation
}
\note{
No documentation XXX all the product of reverse engineering from my part
}
\examples{
\dontrun{
## XXX
}
}
\seealso{
Other AmigaBasic.operations: \code{\link{AmigaBasic.names}},
  \code{\link{AmigaBasic.reserved}},
  \code{\link{[.AmigaBasic}}, \code{\link{as.AmigaBasic}},
  \code{\link{as.character}},
  \code{\link{rawToAmigaBasic}},
  \code{\link{read.AmigaBasic}},
  \code{\link{write.AmigaBasic}}
}
\author{
Pepijn de Vries
}
