% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{AmigaBasic.reserved}
\alias{AmigaBasic.reserved}
\title{List Amiga Basic reserved words.}
\usage{
AmigaBasic.reserved()
}
\value{
Returns a \code{vecor} of \code{character} strings of
reserved Amiga Basic words.
}
\description{
Obtain a list of reserved Amiga Basic words. These words are not
allowed as names of variables or labels in Amiga Basic.
}
\details{
This function will return a full list of reserved Amiga Basic
words. This list does not serve as a manual for basic (for
that purpose consult external resources). This list is meant to
consult when chosing label names in Amiga Basic code. These
reserved words are not allowed as names.
}
\examples{
AmigaBasic.reserved()
}
\seealso{
Other AmigaBasic.operations: \code{\link{AmigaBasic.names}},
  \code{\link{AmigaBasic}}, \code{\link{[.AmigaBasic}},
  \code{\link{as.AmigaBasic}}, \code{\link{as.character}},
  \code{\link{rawToAmigaBasic}},
  \code{\link{read.AmigaBasic}},
  \code{\link{write.AmigaBasic}}
}
\author{
Pepijn de Vries
}
