% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\docType{class}
\name{AmigaBasicShape}
\alias{AmigaBasicShape}
\title{The S3 AmigaBasicShape class}
\description{
XXX implement plot+as.raster function
}
\details{
XXX add to package documentation
}
\examples{
\dontrun{
## XXX
}
}
\seealso{
Other AmigaBasicShape.operations: \code{\link{read.AmigaBasicShape}},
  \code{\link{write.AmigaBasicShape}}
}
\author{
Pepijn de Vries
}
