% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{[.AmigaBasic}
\alias{[.AmigaBasic}
\alias{[<-.AmigaBasic}
\title{Extract or replace lines of Amiga Basic code}
\usage{
\method{[}{AmigaBasic}(x, i)

\method{[}{AmigaBasic}(x, i) <- value
}
\arguments{
\item{x}{An \code{AmigaBasic} class object from which specific lines
need to be extracted or replaced.}

\item{i}{A \code{vector} of \code{numeric} indices. This index
is used to select specific lines. Negative values will deselect lines.}

\item{value}{A \code{vector} of \code{character} strings or an
\code{\link{AmigaBasic}} class object that is used to replace
the selected indices \code{i}. \code{value} should represent the
same number of lines of code as the selected number of lines.}
}
\value{
XXX
}
\description{
XXX
}
\details{
XXX
}
\examples{
\dontrun{
## First generate a few lines of Basic code:
bas <- as.AmigaBasic(c(
  "LET a = 1",
  "a = a + 1",
  "PRINT \\"a now equals\\";a",
  "INPUT \\"clear screen (y/n)? \\", b$",
  "IF b$ = \\"y\\" THEN CLS"
))

## Select only lines 4 and 5:
bas[4:5]

## use negative indices to deselect specific lines.
## deselect line 2:
bas[-2]

## replace line 2
bas[2] <- "a = a + 2"

## You can also use AmigaBasic class object as replacement
bas[2] <- as.AmigaBasic("a = a + 3")
}
}
\seealso{
Other AmigaBasic.operations: \code{\link{AmigaBasic.names}},
  \code{\link{AmigaBasic.reserved}},
  \code{\link{AmigaBasic}}, \code{\link{as.AmigaBasic}},
  \code{\link{as.character}},
  \code{\link{rawToAmigaBasic}},
  \code{\link{read.AmigaBasic}},
  \code{\link{write.AmigaBasic}}
}
\author{
Pepijn de Vries
}
