% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{r_logo.shp}
\alias{r_logo.shp}
\title{TODO}
\format{
See \code{\link{AmigaBasicShape}} for more information
about the format.
}
\description{
TODO
}
\details{
TODO
}
\examples{
\dontrun{
TODO
}
}
