% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{rasterToAmigaBasicShape}
\alias{rasterToAmigaBasicShape}
\title{Convert a grDevices raster object into an AmigaBasicShape class object.}
\usage{
rasterToAmigaBasicShape(x, type = c("blitter object", "sprite"), palette)
}
\arguments{
\item{x}{A \code{\link[grDevices:raster]{raster}} class object to convert into a \code{\link{AmigaBasicShape}} class obejct.}

\item{type}{TODO}

\item{palette}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\details{
TODO
}
\examples{
\dontrun{
## TODO
}
}
\seealso{
Other AmigaBasicShape.operations: 
\code{\link{AmigaBasicShape}},
\code{\link{read.AmigaBasicShape}()},
\code{\link{write.AmigaBasicShape}()}

Other raster.operations: 
\code{\link{AmigaBitmapFont}},
\code{\link{as.raster.AmigaBasicShape}()},
\code{\link{bitmapToRaster}()},
\code{\link{dither}()},
\code{\link{index.colours}()},
\code{\link{rasterToAmigaBitmapFont}()},
\code{\link{rasterToBitmap}()},
\code{\link{rasterToHWSprite}()},
\code{\link{rasterToIFF}()}
}
\author{
Pepijn de Vries
}
\concept{AmigaBasicShape.operations}
\concept{raster.operations}
