% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{rawToAmigaBasic}
\alias{rawToAmigaBasic}
\title{XXX}
\usage{
rawToAmigaBasic(x, ...)
}
\arguments{
\item{x}{A \code{vector} of \code{raw} data that is to be converted
into an \code{\link{AmigaBasic}} class object.}

\item{...}{Currently ignored.}
}
\value{
An \code{\link{AmigaBasic}} class object based on \code{x}.
}
\description{
XXX
}
\details{
XXX
}
\examples{
\dontrun{
## First create an AmigaBAsic object:
bas <- as.AmigaBasic("PRINT \\"Hello world!\\"")

## Make it raw:
bas.raw <- as.raw(bas)

## Now convert it back to an AmigaBasic object:
bas <- rawToAmigaBasic(bas.raw)
}
}
\seealso{
Other AmigaBasic.operations: \code{\link{AmigaBasic.names}},
  \code{\link{AmigaBasic.reserved}},
  \code{\link{AmigaBasic}}, \code{\link{[.AmigaBasic}},
  \code{\link{as.AmigaBasic}}, \code{\link{as.character}},
  \code{\link{read.AmigaBasic}},
  \code{\link{write.AmigaBasic}}

Other raw.operations: \code{\link{as.AmigaBasic}},
  \code{\link{as.raw.AmigaBasic}},
  \code{\link{colourToAmigaRaw}}, \code{\link{packBitmap}},
  \code{\link{rawToAmigaBasicShape}},
  \code{\link{rawToAmigaBitmapFontSet}},
  \code{\link{rawToAmigaBitmapFont}},
  \code{\link{rawToAmigaIcon}},
  \code{\link{rawToHWSprite}}, \code{\link{rawToIFFChunk}},
  \code{\link{rawToSysConfig}},
  \code{\link{simpleAmigaIcon}}
}
\author{
Pepijn de Vries
}
