% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{rawToAmigaBasicShape}
\alias{rawToAmigaBasicShape}
\title{XXX}
\usage{
rawToAmigaBasicShape(x, ...)
}
\arguments{
\item{x}{A \code{vector} of \code{raw} data that is to be converted
into an \code{\link{AmigaBasicShape}} class object.}

\item{...}{Arguments passed on to \code{\link{bitmapToRaster}}.
As the palette is not stored in the file. This can be used to
pass on a custom \code{palette} to \code{\link{bitmapToRaster}}.}
}
\value{
XXX
}
\description{
XXX
}
\details{
XXX
}
\examples{
\dontrun{
## XXX
}
}
\seealso{
Other raw.operations: \code{\link{as.AmigaBasic}},
  \code{\link{as.raw.AmigaBasic}},
  \code{\link{colourToAmigaRaw}}, \code{\link{packBitmap}},
  \code{\link{rawToAmigaBasic}},
  \code{\link{rawToAmigaBitmapFontSet}},
  \code{\link{rawToAmigaBitmapFont}},
  \code{\link{rawToAmigaIcon}},
  \code{\link{rawToHWSprite}}, \code{\link{rawToIFFChunk}},
  \code{\link{rawToSysConfig}},
  \code{\link{simpleAmigaIcon}}
}
\author{
Pepijn de Vries
}
