% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{read.AmigaBasicBMAP}
\alias{read.AmigaBasicBMAP}
\title{Read Amiga Basic BMAP files}
\usage{
read.AmigaBasicBMAP(file, disk = NULL)
}
\arguments{
\item{file}{A \code{character} string of the filename of the Amiga Basic BMAP file to be read.}

\item{disk}{A virtual Commodore Amiga disk from which the \code{file} should be
read. This should be an \code{\link[adfExplorer:amigaDisk-class]{amigaDisk}} object. Using
this argument requires the adfExplorer package.
When set to \code{NULL}, this argument is ignored.}
}
\value{
Returns an \code{\link{AmigaBasicBMAP}} class object read from the \code{file}.
}
\description{
TODO
}
\details{
TODO
}
\examples{
\dontrun{
## TODO
}
}
\seealso{
Other AmigaBasic.operations: 
\code{\link{AmigaBasic.reserved}()},
\code{\link{AmigaBasicBMAP}},
\code{\link{AmigaBasic}},
\code{\link{[.AmigaBasic}()},
\code{\link{as.AmigaBasic}()},
\code{\link{as.character}()},
\code{\link{check.names.AmigaBasic}()},
\code{\link{names.AmigaBasic}()},
\code{\link{rawToAmigaBasicBMAP}()},
\code{\link{rawToAmigaBasic}()},
\code{\link{read.AmigaBasic}()},
\code{\link{write.AmigaBasicBMAP}()},
\code{\link{write.AmigaBasic}()}

Other io.operations: 
\code{\link{read.AmigaBasicShape}()},
\code{\link{read.AmigaBasic}()},
\code{\link{read.AmigaBitmapFontSet}()},
\code{\link{read.AmigaBitmapFont}()},
\code{\link{read.AmigaIcon}()},
\code{\link{read.SysConfig}()},
\code{\link{read.iff}()},
\code{\link{write.AmigaBasicBMAP}()},
\code{\link{write.AmigaBasicShape}()},
\code{\link{write.AmigaBasic}()},
\code{\link{write.AmigaBitmapFont}()},
\code{\link{write.AmigaIcon}()},
\code{\link{write.SysConfig}()},
\code{\link{write.iff}()}
}
\author{
Pepijn de Vries
}
\concept{AmigaBasic.operations}
\concept{io.operations}
