% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{read.AmigaBasicShape}
\alias{read.AmigaBasicShape}
\title{Read Amiga Basic Shape files}
\usage{
read.AmigaBasicShape(file, disk = NULL, ...)
}
\arguments{
\item{file}{A \code{character} string of the filename of the Amiga Basic Shape file to be read.}

\item{disk}{A virtual Commodore Amiga disk from which the \code{file} should be
read. This should be an \code{\link[adfExplorer]{amigaDisk}} object. Using
this argument requires the adfExplorer package.
When set to \code{NULL}, this argument is ignored.}

\item{...}{XXX}
}
\value{
Returns an \code{\link{AmigaBasicShape}} class object read from the \code{file}.
}
\description{
XXX
}
\details{
XXX
}
\examples{
\dontrun{
## XXX
}
}
\seealso{
Other AmigaBasicShape.operations: \code{\link{AmigaBasicShape}},
  \code{\link{write.AmigaBasicShape}}

Other io.operations: \code{\link{read.AmigaBasic}},
  \code{\link{read.AmigaBitmapFontSet}},
  \code{\link{read.AmigaBitmapFont}},
  \code{\link{read.AmigaIcon}},
  \code{\link{read.SysConfig}}, \code{\link{read.iff}},
  \code{\link{write.AmigaBasicShape}},
  \code{\link{write.AmigaBasic}},
  \code{\link{write.AmigaBitmapFont}},
  \code{\link{write.AmigaIcon}},
  \code{\link{write.SysConfig}}, \code{\link{write.iff}}
}
\author{
Pepijn de Vries
}
