% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iff.0.r
\name{read.iff}
\alias{read.iff}
\title{Read Interchange File Format (IFF)}
\usage{
read.iff(file, disk = NULL)
}
\arguments{
\item{file}{A filename of an IFF file to be read, or a connection from which
binary data can be read.}

\item{disk}{A virtual Commodore Amiga disk from which the \code{file} should be
read. This should be an \code{\link[adfExplorer:amigaDisk-class]{amigaDisk()}} object. Using
this argument requires the adfExplorer package.
When set to \code{NULL}, this argument is ignored.}
}
\value{
Returns a \code{\link[=IFFChunk]{IFFChunk()}} object read from the specified file.
}
\description{
Read the Interchange File Format (IFF) as an \code{\link[=IFFChunk]{IFFChunk()}} object.
}
\details{
Information is stored as `chunks' in IFF files (see \code{\link[=IFFChunk]{IFFChunk()}}).
Each chunk should at least contain a label of the type of chunk and the data
for that chunk. This function reads all chunks from a valid IFF file, including
all nested chunks and stores them in an \code{\link[=IFFChunk]{IFFChunk()}} object. IFF
files can hold any kind of data (e.g. images or audio), this read function
does not interpret the file. Use \code{\link[=interpretIFFChunk]{interpretIFFChunk()}} for that
purpose.
}
\examples{
\dontrun{
## let's read a bitmap image stored in IFF as provided with this package:
filename <- system.file("ilbm8lores.iff", package = "AmigaFFH")
example.iff <- read.iff(filename)

## And plot it:
plot(example.iff)
}
}
\seealso{
Other io.operations: 
\code{\link{read.AmigaBasicBMAP}()},
\code{\link{read.AmigaBasicShape}()},
\code{\link{read.AmigaBasic}()},
\code{\link{read.AmigaBitmapFontSet}()},
\code{\link{read.AmigaBitmapFont}()},
\code{\link{read.AmigaIcon}()},
\code{\link{read.SysConfig}()},
\code{\link{write.AmigaBasicShape}()},
\code{\link{write.AmigaBasic}()},
\code{\link{write.AmigaBitmapFont}()},
\code{\link{write.AmigaIcon}()},
\code{\link{write.SysConfig}()},
\code{\link{write.iff}()}

Other iff.operations: 
\code{\link{IFFChunk-class}},
\code{\link{WaveToIFF}()},
\code{\link{as.raster.AmigaBasicShape}()},
\code{\link{getIFFChunk}()},
\code{\link{interpretIFFChunk}()},
\code{\link{rasterToIFF}()},
\code{\link{rawToIFFChunk}()},
\code{\link{write.iff}()}
}
\author{
Pepijn de Vries
}
\concept{iff.operations}
\concept{io.operations}
