% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{write.AmigaBasicBMAP}
\alias{write.AmigaBasicBMAP}
\title{Write an AmigaBasicBMAP object to a file}
\usage{
write.AmigaBasicBMAP(x, file, disk = NULL)
}
\arguments{
\item{x}{The \code{\link{AmigaBasicBMAP}} class object that needs to be
stored.}

\item{file}{A \code{character} string specifying the file location
to which \code{x} (an \code{\link{AmigaBasicBMAP}} object) needs to be written.}

\item{disk}{A virtual Commodore Amiga disk to which the \code{file} should be
written. This should be an \code{\link[adfExplorer:amigaDisk-class]{amigaDisk}} object. Using
this argument requires the adfExplorer package.
When set to \code{NULL}, this argument is ignored.}
}
\value{
Invisibly returns the result of the call of \code{close} to the
file connection. Or, when \code{disk} is specified, a copy of
\code{disk} is returned to which the file(s) is/are written.
}
\description{
Write an \code{\link{AmigaBasicBMAP}} class object to a file in its binary format.
}
\details{
TODO details
}
\examples{
\dontrun{
## TODO
}
}
\seealso{
Other AmigaBasic.operations: 
\code{\link{AmigaBasic.reserved}()},
\code{\link{AmigaBasicBMAP}},
\code{\link{AmigaBasic}},
\code{\link{[.AmigaBasic}()},
\code{\link{as.AmigaBasicBMAP}()},
\code{\link{as.AmigaBasic}()},
\code{\link{as.character}()},
\code{\link{check.names.AmigaBasic}()},
\code{\link{names.AmigaBasic}()},
\code{\link{rawToAmigaBasicBMAP}()},
\code{\link{rawToAmigaBasic}()},
\code{\link{read.AmigaBasicBMAP}()},
\code{\link{read.AmigaBasic}()},
\code{\link{write.AmigaBasic}()}

Other io.operations: 
\code{\link{read.AmigaBasicBMAP}()},
\code{\link{read.AmigaBasicShape}()},
\code{\link{read.AmigaBasic}()},
\code{\link{read.AmigaBitmapFontSet}()},
\code{\link{read.AmigaBitmapFont}()},
\code{\link{read.AmigaIcon}()},
\code{\link{read.SysConfig}()},
\code{\link{read.iff}()},
\code{\link{write.AmigaBasicShape}()},
\code{\link{write.AmigaBasic}()},
\code{\link{write.AmigaBitmapFont}()},
\code{\link{write.AmigaIcon}()},
\code{\link{write.SysConfig}()},
\code{\link{write.iff}()}
}
\author{
Pepijn de Vries
}
\concept{AmigaBasic.operations}
\concept{io.operations}
