% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardwareSprites.r, R/iff.0.r, R/iff.8svx.r,
%   R/iff.ilbm.r
\name{plot}
\alias{plot}
\alias{plot.hardwareSprite}
\alias{plot}
\alias{plot.IFFChunk}
\alias{plot}
\alias{plot.IFF.FORM}
\alias{plot}
\alias{plot.IFF.8SVX}
\alias{plot}
\alias{plot.IFF.ILBM}
\alias{plot}
\alias{plot.IFF.ANIM}
\title{Plot AmigaFFH objects}
\usage{
\method{plot}{hardwareSprite}(x, y, ...)

\method{plot}{IFFChunk}(x, y, ...)

\method{plot}{IFF.FORM}(x, y, ...)

\method{plot}{IFF.8SVX}(x, y, ...)

\method{plot}{IFF.ILBM}(x, y, ...)

\method{plot}{IFF.ANIM}(x, y, ...)
}
\arguments{
\item{x}{An AmigaFFH object to be plotted. See usage section for supported object
classes.}

\item{y}{This argument is inherited from the generic plot-function but is currently
ignored in this package.}

\item{...}{Parameters passed onto the generic \code{graphics} plotting routine.}
}
\value{
Returns \code{NULL} silently.
}
\description{
Plot AmigaFFH objects using \code{base} plotting routines.
}
\details{
A plotting routine is implemented for most AmigaFFH objects. See the usage section
for all supported objects.
}
\examples{
\dontrun{
## load an IFF file
example.iff <- read.iff(system.file("ilbm8lores.iff", package = "AmigaFFH"))

## and plot it:
plot(example.iff)
}
}
\author{
Pepijn de Vries
}
