% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.r
\name{rasterToAmigaBasicShape}
\alias{rasterToAmigaBasicShape}
\title{Convert a grDevices raster object into an AmigaBasicShape class object.}
\usage{
rasterToAmigaBasicShape(x, type = c("blitter object", "sprite"), palette, ...)
}
\arguments{
\item{x}{A \code{\link[grDevices:as.raster]{raster}} class object to convert into a \code{\link{AmigaBasicShape}} class obejct.}

\item{type}{A \code{character} string indicating what type of graphic needs to be created: "\code{blitter object}" (default) or "\code{sprite}".}

\item{palette}{A \code{vector} of \code{character} strings, where each element represents a colour. This palette is used to quantize the
colours that occur in the \code{raster} \code{x}.}

\item{...}{Arguments passed onto \code{\link{index.colours}}. Can be used, for instance, to achieve specific dithering effects.}
}
\value{
Returns an \code{\link{AmigaBasicShape}} class object based on \code{x}.
}
\description{
Convert a \code{\link[grDevices:as.raster]{raster}} object into an \code{\link{AmigaBasicShape}} class object.
}
\details{
TODO
}
\examples{
\dontrun{
## get a raster image:
ilbm <- as.raster(read.iff(system.file("ilbm8lores.iff", package = "AmigaFFH")))

## convert to an Amiga Basic blitter object:
bob <- rasterToAmigaBasicShape(ilbm, "blitter object")
}
}
\seealso{
Other AmigaBasicShape.operations: 
\code{\link{AmigaBasicShape}},
\code{\link{read.AmigaBasicShape}()},
\code{\link{write.AmigaBasicShape}()}

Other raster.operations: 
\code{\link{AmigaBitmapFont}},
\code{\link{as.raster.AmigaBasicShape}()},
\code{\link{bitmapToRaster}()},
\code{\link{dither}()},
\code{\link{index.colours}()},
\code{\link{rasterToAmigaBitmapFont}()},
\code{\link{rasterToBitmap}()},
\code{\link{rasterToHWSprite}()},
\code{\link{rasterToIFF}()}
}
\author{
Pepijn de Vries
}
\concept{AmigaBasicShape.operations}
\concept{raster.operations}
